Ignore changes to aliases
This commit is not intended ever to go to "master" branch. The only reason
it exists is because without a pre-terraform-run templating it is not possible
to customize the lifecycle section and we don't want to do that for just this
single use-case. Terragrunt are a bit too invasive for the job, make
assumptions about the code layout and fall short by not being able to inject
code into existing files.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.net>